requirements.txt
apex==0.9.10dev
attrdict==2.0.1
einops==0.6.0
elf==0.4.5
GPUtil==1.4.0
h5py==3.8.0
joblib==1.2.0
mahotas==1.4.13
matplotlib==3.5.1
monai==1.1.0
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.5.0
PyYAML==6.0
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.7.3
SimpleITK==2.2.1
skimage==0.0
tensorboardX==2.6
torch==1.12.0
torchio==0.18.87
torchvision==0.13.0
tqdm==4.64.1
waterz==0.9.4
